np.int to np.int64 np.bool to np.bool_

+++ src/qudi/hardware/awg/keysight_m819x.py
@@ -1320,7 +1320,7 @@ class AWGM819X(PulserInterface):
                 if name.split(',')[0] == name:
                     segment_id = self.asset_name_2_id(name, chnl_num, mode='segment')
                 else:
-                    segment_id = np.int(name.split(',')[0])
+                    segment_id = np.int64(name.split(',')[0])
                 self.write(':TRAC{0}:SEL {1}'.format(chnl_num, segment_id))
 
 
@@ -1546,7 +1546,7 @@ class AWGM819X(PulserInterface):
                 segment_id = to_segment_id
                 if name.split(',')[0] != name:
                     # todo: this breaks if there is a , in the name without number
-                    segment_id = np.int(name.split(',')[0])
+                    segment_id = np.int64(name.split(',')[0])
                     self.log.warning("Loading wave to specified segment ({}) via name will deprecate.".format(segment_id))
                 if segment_id == -1:
                     # to next free segment
